+++ Classes/Domain/Model/Info.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace FriendsOfTYPO3\BlogExample\Domain\Model;
+namespace T3docs\BlogExample\Domain\Model;
 
 /*
  * This file is part of the TYPO3 CMS project.